+++ oshi-core/src/main/java/oshi/util/platform/windows/WmiQueryHandler.java
@@ -39,12 +39,22 @@ import com.sun.jna.platform.win32.COM.COMUtils;
 import com.sun.jna.platform.win32.COM.Wbemcli;
 import com.sun.jna.platform.win32.COM.WbemcliUtil;
 
+import oshi.util.GlobalConfig;
+
 public class WmiQueryHandler {
 
     private static final Logger LOG = LoggerFactory.getLogger(WmiQueryHandler.class);
 
+    private static int globalTimeout = GlobalConfig.get("oshi.util.wmi.timeout", -1);
+
+    static {
+        if (globalTimeout == 0 || globalTimeout < -1) {
+            throw new GlobalConfig.PropertyException("oshi.util.wmi.timeout");
+        }
+    }
+
     // Timeout for WMI queries
-    private int wmiTimeout = Wbemcli.WBEM_INFINITE;
+    private int wmiTimeout = globalTimeout;
 
     // Cache failed wmi classes
     private final Set<String> failedWmiClassNames = new HashSet<>();